# McShares.postman_collection.json-OpenApi3Json.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "McShares",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "variables": {}
    }
  ],
  "paths": {
...
    "/mcShares/uploadAndValidateSharesFile": {
      "post": {
        "tags": [
          "Shares/Customers"
        ],
        "summary": "uploadAndValidateSharesFile",
        "operationId": "uploadAndValidateSharesFile",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "encoding": {},
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "description": "Upload, validate and persist customer shares data from xml file",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
    "/mcShares/updateSharesDetails/C11233": {
      "put": {
        "tags": [
          "Shares/Customers"
        ],
        "summary": "updateSharesDetails",
        "operationId": "updateSharesDetails",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/updateSharesDetailsRequest"
              },
              "example": {
                "numShares": 100,
                "sharePrice": 5
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "update of specific records of shares",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
  },
  "components": {
    "schemas": {
      "updateSharesDetailsRequest": {
        "title": "updateSharesDetailsRequest",
        "required": [
          "numShares",
          "sharePrice"
        ],
        "type": "object",
        "properties": {
          "numShares": {
            "type": "integer",
            "format": "int32"
          },
          "sharePrice": {
            "type": "integer",
            "format": "int32"
          }
        },
        "example": {
          "numShares": 100,
          "sharePrice": 5
        }
      },
...
    },
    "securitySchemes": {
      "httpBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "httpBearer": []
    }
  ],
...
}